feat: Added mod cards component
