+++ desktop/src/ui/pages/browser/BrowserMain.tsx
@@ -1,10 +1,9 @@
-import { Download01Icon, FileCode01Icon, FilterLinesIcon, HeartIcon, SearchMdIcon } from '@untitled-theme/icons-solid';
-import { type Accessor, For, type Setter, createEffect, createSignal } from 'solid-js';
+import { Download01Icon, FileCode01Icon, FilterLinesIcon, SearchMdIcon } from '@untitled-theme/icons-solid';
+import { For, createSignal } from 'solid-js';
 import Button from '~ui/components/base/Button';
 import Dropdown from '~ui/components/base/Dropdown';
 import TextField from '~ui/components/base/TextField';
-import ScrollableContainer from '~ui/components/ScrollableContainer';
-import { abbreviateNumber } from '~utils/primitives';
+import ModCard, { Provider } from '~ui/components/content/ModCard';
 import createSortable from '~utils/sorting';
 
 interface CardProps {
@@ -14,13 +13,6 @@ interface CardProps {
 	date_updated: number;
 	provider: Provider;
 };
-
-enum Provider {
-	Curseforge,
-	Modrinth,
-	Polyfrost,
-	Skyclient,
-}
 
 interface BrowserFilters {
 	provider: Provider[]; // TODO: Change with Provider types from Rust
@@ -90,22 +82,25 @@ function BrowserMain() {
 	);
 }
 
+export const modStore: Mod[] = [
+	{
+		id: '1yIQcc2b',
+		name: 'EvergreenHUD',
+		description: 'Improves your heads up display.',
+		author: 'Polyfrost',
+		icon_url: 'https://cdn.modrinth.com/data/1yIQcc2b/icon.png',
+		page_url: 'https://modrinth.com/mod/evergreenhud',
+		provider: Provider.Modrinth,
+		downloads: 281700,
+		ratings: 220,
+	},
+];
+
 const _modRowData: ModsRowProps[] = [
 	{
 		header: 'Highly Endorsed',
 		category: 'endorsed',
-		mods: [
-			{
-				name: 'EvergreenHUD',
-				description: 'Improves your heads up display.',
-				author: 'Polyfrost',
-				icon_url: 'https://cdn.modrinth.com/data/1yIQcc2b/icon.png',
-				page_url: 'https://modrinth.com/mod/evergreenhud',
-				provider: Provider.Modrinth,
-				downloads: 281700,
-				ratings: 220,
-			},
-		],
+		mods: modStore,
 	},
 ];
 
@@ -114,17 +109,6 @@ interface ModsRowProps {
 	category: string;
 	mods: Mod[];
 };
-
-interface Mod {
-	name: string;
-	description: string;
-	author: string;
-	icon_url: string;
-	provider: Provider;
-	page_url: string;
-	downloads: number;
-	ratings: number;
-}
 
 function ModsRow(props: ModsRowProps) {
 	return (
@@ -140,45 +124,6 @@ function ModsRow(props: ModsRowProps) {
 				</For>
 			</div>
 
-		</div>
-	);
-}
-
-function ModCard(props: Mod) {
-	return (
-		<div class="flex flex-col overflow-hidden rounded-lg bg-component-bg max-w-53 min-w-53 w-full max-h-68 min-h-68 h-full">
-			<div class="relative h-28 overflow-hidden flex justify-center items-center">
-				<img class="absolute filter-blur-xl z-0 max-w-none w-7/6" src={props.icon_url} alt={`Icon for ${props.name}`} />
-				<img class="relative w-2/5 aspect-ratio-square z-1 rounded-2xl" src={props.icon_url} alt={`Icon for ${props.name}`} />
-			</div>
-			<div class="flex flex-col flex-1 p-3 gap-4">
-				<div class="flex flex-col gap-2">
-					<h5 class="font-medium text-fg-primary">{props.name}</h5>
-					<p class="text-fg-secondary text-xs">
-						By
-						{' '}
-						<span class="text-fg-primary">{props.author}</span>
-						{' '}
-						on
-						{' '}
-						{Provider[props.provider]}
-					</p>
-				</div>
-
-				<p class="text-fg-secondary text-sm flex-1">{props.description}</p>
-
-				<div class="flex flex-row gap-4 text-xs">
-					<div class="flex flex-row items-center gap-2">
-						<Download01Icon class="w-4 h-4" />
-						{abbreviateNumber(props.downloads)}
-					</div>
-
-					<div class="flex flex-row items-center gap-2">
-						<HeartIcon class="w-4 h-4" />
-						{abbreviateNumber(props.ratings)}
-					</div>
-				</div>
-			</div>
 		</div>
 	);
 }
